+++ ai/backend/integration/jupyter/install.py
@@ -1,7 +1,7 @@
 '''
 The kernel installer.
 
-Run `python -m sorna.integration.jupyter.install` to use Sorna in your Jupyter notebooks.
+Run `python -m ai.backend.integration.jupyter.install` to use Backend.AI in your Jupyter notebooks.
 '''
 
 import argparse
@@ -12,7 +12,7 @@ import webbrowser
 
 from jupyter_client.kernelspec import KernelSpecManager
 from IPython.utils.tempdir import TemporaryDirectory
-from .kernel import sorna_kernels
+from .kernel import kernels
 
 
 def clean_kernel_spec(user=True, prefix=None):
@@ -21,8 +21,8 @@ def clean_kernel_spec(user=True, prefix=None):
     #       Sometimes we may need to perform --clean-only multiple times to completely
     #       remove all kernelspecs installed around venvs and system global directories.
     for name, info in mgr.get_all_specs().items():
-        if name.startswith('sorna'):
-            print(f"Removing existing Sorna kernel: {info['spec']['display_name']}")
+        if name.startswith('backend'):
+            print(f"Removing existing Backend.AI kernel: {info['spec']['display_name']}")
             mgr.remove_kernel_spec(name)
 
 
@@ -31,7 +31,7 @@ def install_kernel_spec(name, spec_json, user=True, prefix=None):
         os.chmod(td, 0o755) # Starts off as 700, not user readable
         with open(os.path.join(td, 'kernel.json'), 'w') as f:
             json.dump(spec_json, f, sort_keys=True)
-        print(f"Installing Sorna-backed Jupyter kernel spec: {spec_json['display_name']}")
+        print(f"Installing Backend.AI Jupyter kernel spec: {spec_json['display_name']}")
         KernelSpecManager().install_kernel_spec(
             td, name, user=user, replace=True, prefix=prefix)
 
@@ -60,7 +60,7 @@ def main(argv=None):
     ap.add_argument('--sys-prefix', action='store_true',
         help="Install to sys.prefix (e.g. a virtualenv or conda env)")
     ap.add_argument('--clean-only', action='store_true',
-        help="Perform only clean-up of existing Sorna kernels.")
+        help="Perform only clean-up of existing Backend.AI kernels.")
     ap.add_argument('-q', '--quiet', action='store_true',
         help="Do not ask the user anything.")
     ap.add_argument('--prefix',
@@ -77,9 +77,9 @@ def main(argv=None):
     if args.clean_only:
         return
 
-    for kern in sorna_kernels:
+    for kern in kernels:
         spec = {
-            "argv": [sys.executable, "-m", "sorna.integration.jupyter",
+            "argv": [sys.executable, "-m", "ai.backend.integration.jupyter",
                      "-f", "{connection_file}",
                      "--",
                      "-k", kern.__name__],
@@ -90,7 +90,7 @@ def main(argv=None):
 
     if not args.quiet:
         print()
-        has_api_key = bool(os.environ.get('SORNA_ACCESS_KEY', ''))
+        has_api_key = bool(os.environ.get('BACKEND_ACCESS_KEY', ''))
         if has_api_key:
             print('It seems that you already configured the API key. Enjoy!')
         else:
@@ -99,8 +99,8 @@ def main(argv=None):
             print()
             print('If you already have the keypair or just grabbed a new one,')
             print('run the following in your shell before running jupyter notebook:\n')
-            print('  export SORNA_ACCESS_KEY="AKIA..."')
-            print('  export SORNA_SECRET_KEY="......."\n')
+            print('  export BACKEND_ACCESS_KEY="AKIA..."')
+            print('  export BACKEND_SECRET_KEY="......."\n')
 
 
 if __name__ == '__main__':